requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
coverage>=7.0.0
tqdm==4.66.1
rich==13.7.0

